+++ src/index.js
@@ -59,9 +59,10 @@ export default function loader(content) {
     this.emitFile(outputPath, content);
   }
 
-  return `${
-    options.esModules ? 'export default' : 'module.exports ='
-  } ${publicPath};`;
+  const esModules =
+    typeof options.esModules !== 'undefined' ? options.esModules : true;
+
+  return `${esModules ? 'export default' : 'module.exports ='} ${publicPath};`;
 }
 
 export const raw = true;