+++ test/fixtures/emitFile/fixture.js
@@ -1,4 +1,0 @@
-/* eslint-disable */
-import svg from '../file.svg';
-
-export default svg