+++ test/fixtures/nested/fixture.js
@@ -1,4 +1,0 @@
-/* eslint-disable */
-import png from './assets/file.png';
-
-export default png;