+++ test/fixtures/simple.js
@@ -1,4 +1,6 @@
 /* eslint-disable */
 import png from './file.png';
 
+__export__ = png;
+
 export default png;