+++ test/helpers/index.js
@@ -1,0 +1,7 @@
+import compile from './compile';
+import execute from './execute';
+import getCompiler from './getCompiler';
+import normalizeErrors from './normalizeErrors';
+import readAsset from './readAsset';
+
+export { compile, execute, getCompiler, normalizeErrors, readAsset };